// src/generated/client/models/ExhibitionPageLayoutView.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Muisti API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import {
    ExhibitionPageLayoutViewProperty,
    ExhibitionPageLayoutViewPropertyFromJSON,
    ExhibitionPageLayoutViewPropertyFromJSONTyped,
    ExhibitionPageLayoutViewPropertyToJSON,
} from './';

/**
 * 
 * @export
 * @interface ExhibitionPageLayoutView
 */
export interface ExhibitionPageLayoutView {
    /**
     * 
     * @type {string}
     * @memberof ExhibitionPageLayoutView
     */
    id: string;
    /**
     * 
     * @type {string}
     * @memberof ExhibitionPageLayoutView
     */
    widget: string;
    /**
     * 
     * @type {Array<ExhibitionPageLayoutViewProperty>}
     * @memberof ExhibitionPageLayoutView
     */
    properties: Array<ExhibitionPageLayoutViewProperty>;
    /**
     * 
     * @type {Array<ExhibitionPageLayoutView>}
     * @memberof ExhibitionPageLayoutView
     */
    children: Array<ExhibitionPageLayoutView>;
}

export function ExhibitionPageLayoutViewFromJSON(json: any): ExhibitionPageLayoutView {
    return ExhibitionPageLayoutViewFromJSONTyped(json, false);
}

export function ExhibitionPageLayoutViewFromJSONTyped(json: any, ignoreDiscriminator: boolean): ExhibitionPageLayoutView {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'id': json['id'],
        'widget': json['widget'],
        'properties': ((json['properties'] as Array<any>).map(ExhibitionPageLayoutViewPropertyFromJSON)),
        'children': ((json['children'] as Array<any>).map(ExhibitionPageLayoutViewFromJSON)),
    };
}

export function ExhibitionPageLayoutViewToJSON(value?: ExhibitionPageLayoutView | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'id': value.id,
        'widget': value.widget,
        'properties': ((value.properties as Array<any>).map(ExhibitionPageLayoutViewPropertyToJSON)),
        'children': ((value.children as Array<any>).map(ExhibitionPageLayoutViewToJSON)),
    };
}