// src/generated/client/models/ExhibitionPageLayoutViewProperty.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Muisti API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import {
    ExhibitionPageLayoutViewPropertyType,
    ExhibitionPageLayoutViewPropertyTypeFromJSON,
    ExhibitionPageLayoutViewPropertyTypeFromJSONTyped,
    ExhibitionPageLayoutViewPropertyTypeToJSON,
} from './';

/**
 * 
 * @export
 * @interface ExhibitionPageLayoutViewProperty
 */
export interface ExhibitionPageLayoutViewProperty {
    /**
     * 
     * @type {string}
     * @memberof ExhibitionPageLayoutViewProperty
     */
    name: string;
    /**
     * 
     * @type {string}
     * @memberof ExhibitionPageLayoutViewProperty
     */
    value: string;
    /**
     * 
     * @type {ExhibitionPageLayoutViewPropertyType}
     * @memberof ExhibitionPageLayoutViewProperty
     */
    type: ExhibitionPageLayoutViewPropertyType;
}

export function ExhibitionPageLayoutViewPropertyFromJSON(json: any): ExhibitionPageLayoutViewProperty {
    return ExhibitionPageLayoutViewPropertyFromJSONTyped(json, false);
}

export function ExhibitionPageLayoutViewPropertyFromJSONTyped(json: any, ignoreDiscriminator: boolean): ExhibitionPageLayoutViewProperty {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'name': json['name'],
        'value': json['value'],
        'type': ExhibitionPageLayoutViewPropertyTypeFromJSON(json['type']),
    };
}

export function ExhibitionPageLayoutViewPropertyToJSON(value?: ExhibitionPageLayoutViewProperty | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'name': value.name,
        'value': value.value,
        'type': ExhibitionPageLayoutViewPropertyTypeToJSON(value.type),
    };
}