// src/generated/client/models/ExhibitionPageLayoutViewPropertyType.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Muisti API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 * 
 * @export
 * @enum {string}
 */
export enum ExhibitionPageLayoutViewPropertyType {
    String = 'string',
    Boolean = 'boolean',
    Number = 'number'
}

export function ExhibitionPageLayoutViewPropertyTypeFromJSON(json: any): ExhibitionPageLayoutViewPropertyType {
    return ExhibitionPageLayoutViewPropertyTypeFromJSONTyped(json, false);
}

export function ExhibitionPageLayoutViewPropertyTypeFromJSONTyped(json: any, ignoreDiscriminator: boolean): ExhibitionPageLayoutViewPropertyType {
    return json as ExhibitionPageLayoutViewPropertyType;
}

export function ExhibitionPageLayoutViewPropertyTypeToJSON(value?: ExhibitionPageLayoutViewPropertyType | null): any {
    return value as any;
}